fixed a problem where not having a .lateslugignore would cause the buildpack to fail

+++ lib/language_pack/ruby.rb
@@ -628,44 +628,44 @@ params = CGI.parse(uri.query || "")
   
   def late_slug_ignore
     # Meh, I should log something here.
-    topic("Beep Bloop. Processing your .lateslugignore file!.")
-    ignored_extensions = Array.new
-
-    late_slug_ignore_file = File.new(".lateslugignore", "r")
-    late_slug_ignore_file.each do |line|
-      unless line.chomp!.empty?
-        ignored_extensions.push line
+    if File.exist?(".lateslugignore")
+      topic("Beep Bloop. Processing your .lateslugignore file!.")
+      ignored_extensions = Array.new
+  
+      late_slug_ignore_file = File.new(".lateslugignore", "r")
+      late_slug_ignore_file.each do |line|
+        unless line.chomp!.empty?
+          ignored_extensions.push line
+        end
       end
-    end
-    late_slug_ignore_file.close
+      late_slug_ignore_file.close
 
-    matched_files = Array.new
-    ignored_extensions.each {|ext| matched_files.push Dir.glob(File.join("**",ext))}
-    matched_files.flatten!
-    puts "Deleting #{matched_files.count} files matching .lateslugignore patterns."
-    matched_files.each { |f| File.delete(f)}
+      matched_files = Array.new
+      ignored_extensions.each {|ext| matched_files.push Dir.glob(File.join("**",ext))}
+      matched_files.flatten!
+      puts "Deleting #{matched_files.count} files matching .lateslugignore patterns."
+      matched_files.each { |f| File.delete(f)}
 
-    # For what it's worth, I wrote an asset cleaning tool, but it's not generic enough for general use, but I bet
-    # it probably does a better job achieving a completely clean asset configuration when used in tandem with
-    # asset_sync -- then again, I've only lightly considered this.  Anyway, if someone cares to improve this, the
-    # code is sitting right below
-    #
-    #puts "Running rake assets:clean"
-    #require 'benchmark'
-    #time = Benchmark.realtime { pipe("env PATH=$PATH:bin bundle exec rake assets:clean 2>&1") }
-    #if $?.success?
-    #  # Really, for the love of god, why does the string formatting look so crazy???
-    #  puts "Assets cleaned from compilation location in (#{"%.2f" % time}s)."
-    #else
-    #  puts "Asset cleansing failed.  Yikes."
-    #end
-    #puts "Dropping assets from app/assets, lib/assets, and vendor/assets."
-    #FileUtils.rm_rf("app/assets")
-    #FileUtils.rm_rf("lib/assets")
-    #FileUtils.rm_rf("vendor/assets")
-    #puts "All assets removed from the slug."
-
-
-
+      # For what it's worth, I wrote an asset cleaning tool, but it's not generic enough for general use, but I bet
+      # it probably does a better job achieving a completely clean asset configuration when used in tandem with
+      # asset_sync -- then again, I've only lightly considered this.  Anyway, if someone cares to improve this, the
+      # code is sitting right below
+      #
+      #puts "Running rake assets:clean"
+      #require 'benchmark'
+      #time = Benchmark.realtime { pipe("env PATH=$PATH:bin bundle exec rake assets:clean 2>&1") }
+      #if $?.success?
+      #  # Really, for the love of god, why does the string formatting look so crazy???
+      #  puts "Assets cleaned from compilation location in (#{"%.2f" % time}s)."
+      #else
+      #  puts "Asset cleansing failed.  Yikes."
+      #end
+      #puts "Dropping assets from app/assets, lib/assets, and vendor/assets."
+      #FileUtils.rm_rf("app/assets")
+      #FileUtils.rm_rf("lib/assets")
+      #FileUtils.rm_rf("vendor/assets")
+      #puts "All assets removed from the slug."
+  else
+     topic("Beep Bloop. Failed to find your .lateslugignore file!.  Is it in your applications root directory?")
   end
 end
